Add files via upload
Your dataset should be called "CAD" and should be in csv format in which each row is a unique patient. The column denoting the outcome should be named as described in the script where 0 means the patient was censored (lost to follow-up without developing outcome, 1 means patient developed the outcome, and 2 where patient died without developing the outcome (2 is only possible in non-fatal outcomes such as MI, CHF, and stroke, not in all-cause mortality or MACE). {outcome}_TIME should denote time to first occurrence of outcome diagnosis, loss to follow-up, or mortality, whichever came first. If some patients in your dataset have pre-existing {outcome} and you would like to not include them in the analysis, use PRE_{outcome} as a binary variable to denote presence of existing disease.